+++ src/app/requirements.txt
@@ -6,5 +6,6 @@ openai==0.27.0
 pandas==1.5.3
 requests==2.28.2
 scikit-learn==1.1.2
-streamlit==1.19.0
+streamlit==1.22.0
+streamlit-extras==0.2.7
 xgboost==1.6.2
